+++ python/eyes_selenium/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [vNext]
 ### Added
 - Log errors received from SDK server before re-raising them [Trello 2991](https://trello.com/c/PSEPcMyh)
+- Support adding `user_test_id` to eyes.open config step [Trello 2977](https://trello.com/c/lYJnhwCz)
 ### Fixed
 - Cyclic import in applittols.core module [Trello 2985](https://trello.com/c/akCp8wcr)
 ### Updated